requirements.txt
opencv-python~=4.8
numpy>=1.20
watchdog~=3.0
pynetworktables